% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.div.R
\name{alpha.div}
\alias{alpha.div}
\title{Calculate Alpha Diversity Metrics}
\usage{
alpha.div(x, sample.col = 1)
}
\arguments{
\item{x}{Data table with sample IDs in the first column}

\item{sample.col}{Column name containing sample identifiers}
}
\value{
A data frame containing associated alpha diversity metrics for samples in the provided table.
}
\description{
This function is used to calculate multiple alpha diversity metrics all in one go. This calculates Shannon Index, Simpson Index, Pielou Index, and sample richness
}
