% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.qr.R
\name{generate.qr}
\alias{generate.qr}
\title{QR Code Generation}
\usage{
generate.qr(site, filename, path = paste0(getwd(), "/plots/svg"), show = FALSE)
}
\arguments{
\item{site}{URL of the website to generate the QR code for.}

\item{filename}{Name of the saved SVG file containing the QR code. Do not include ".svg" in the name.}

\item{path}{Path where the QR SVG should be saved. Defaults to a "svg" sub-directory within the "plots" directory.}

\item{show}{Determines if the generated SVG is immediately opened. Defaults to FALSE.}
}
\description{
Generates a QR code for a given website URL and saves it as an SVG file.
}
