% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guzioR-package.R, R/palettes.R
\docType{data}
\name{guzioR}
\alias{guzioR}
\alias{guzior.pal}
\alias{b10.pal}
\title{Internal package methods}
\format{
\subsection{\code{guzior.pal}}{

An object of class \code{palettes_palette} with
3 color palettes.
Use \code{names(guzior.pal)} to return all palette names.
}

\subsection{\code{b10.pal}}{

An object of class \code{palettes_palette} with
13 color palettes.
Use \code{names(b10.pal)} to return all B10 palette names.
}
}
\usage{
guzior.pal

b10.pal
}
\description{
These include a colorblind friendly palette set in addition to a
set of palettes based on the brand identities of all 14 Big Ten
schools (as of April 2024).
}
\keyword{datasets}
\keyword{internal}
