% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guzior.setup.R
\name{guzior.setup}
\alias{guzior.setup}
\title{Doug's Typical Directory Setup}
\usage{
guzior.setup()
}
\description{
This sets up sub-directories in a way that I typically use. By no means do you need to do this, simply comes down to preference.
}
