% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guzior_setdir.R
\name{guzior_setdir}
\alias{guzior_setdir}
\title{Doug's Typical Directory Setup}
\usage{
guzior_setdir()
}
\description{
This sets up sub-directories in a way that I typically use. By no means do you need to do this, simply comes down to preference.
}
