% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_facet.R
\name{not_facet}
\alias{not_facet}
\title{Plot Facets as Single Plot}
\usage{
not_facet(x)
}
\arguments{
\item{x}{GGplot object}
}
\description{
Add to the end of ggplot code to shrink space between facets, move the strip label, and plot as if objects are grouped by the facets themselves.
}
