% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielou.R
\name{pielou}
\alias{pielou}
\title{Calculate Pielou Evenness Index}
\usage{
pielou(x)
}
\arguments{
\item{x}{Data containing only numeric values, with samples in rows and features in columns}
}
\value{
A data vector containing calculated index values
}
\description{
This function is used to calculate how evenly features are distributed per sample across a data frame or data table.
}
