% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.gen.R
\name{qr.gen}
\alias{qr.gen}
\title{QR Code Generation}
\usage{
\method{qr}{gen}(site, filename)
}
\arguments{
\item{site}{Name of the address you want to make the QR code for.}

\item{path}{Name of the saved SVG file containing the QR code. Do not include ".svg" here.}
}
\description{
Paste the website you want to make a QR code for and save it. QR code generator comes from the package 'qrcode'.
}
