% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.ggplot.R
\name{save.ggplot}
\alias{save.ggplot}
\title{Save Plot as Multiple File Types}
\usage{
save.ggplot(plot, path, filename, height, width)
}
\arguments{
\item{plot}{Name of the plot object to be saved}

\item{path}{Path to the designated plot folder. These is where subfolders will be built for each plot file type.}

\item{filename}{What you want the saved plot to be named}

\item{height}{Plot height}

\item{width}{Plot width}
}
\description{
This function saves the specified plots built using ggplot2 as png, pdf, and svg files. You can specify subfolders for each data subset or store plots however you want in the plot directory.
}
