% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pheatmap.R
\name{save_pheatmap}
\alias{save_pheatmap}
\title{Save pheatmap function (png or pdf)}
\usage{
save_pheatmap(x, filename, width = 12, height = 12)
}
\arguments{
\item{x}{pheatmap object}

\item{filename}{Desired filename, containing png or pdf}

\item{width}{Plot width for file}

\item{height}{Plot height for file}
}
\description{
Function to save the heat map generated by pheatmap, as first written by Matt
Whitaker (https://gist.github.com/mathzero)
}
\details{
function detects '.png' or '.pdf' in the declared filename and assigns that
file type
}
