% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{b10.pal}
\alias{b10.pal}
\alias{guzior.pal}
\title{Commonly used palettes}
\format{
\subsection{\code{b10.pal}}{

An object of class \code{palettes_palette} with
13 discrete colour palettes.
Use \code{names(b10.pal)} to return all palette names.
}

\subsection{\code{guzior.pal}}{

An object of class \code{palettes_palette} with
2 sequential colour palettes.
Use \code{names(guzior.pal)} to return all palette names.
}
}
\usage{
b10.pal

guzior.pal
}
\description{
Commonly used palettes
}
\keyword{datasets}
